Add support for bundling requests
Since we are notified by different events for instances that have
same request_spec, it is better to bundle the requests and try to
free resources for all the servers at once. At the same time we need
to make sure that the reaper will make space only for non-preemptible
requests.