Reaper as an independent service
The reaper should be a new service dependent only on its jobboard.
For this reason we remove reaper's dependencies from the job_manager.
The job_manager is only used as an interface for other services, to
post jobs on the reaper's jobboard.

Additionally we add a new entry point responsible for starting up the
new reaper service.