Integrate state calculation with Taskflow jobs
If the new config option aardvark.enable_watermark_mode is set to True
then a periodic task is added to the reaper_service that posts
StateCalculation jobs on the reaper's jobboard.

For this reason we remove the entry point for the old state calculation
service since it's not needed anymore.