Enhance the worker health check
Every time the periodic task to check the health of the workers is
executed, it will increase the worker's missed_acks by 1. If
missed_acks go up to 3 then, then we assume the thread is dead, we stop
it and we join with it (interval 0.1). Then a new worker is spawned
replacing the dead one.

Each worker, in order to show that it's alive, resets the missed_acks
to 0 in every loop.