Requeue notification if scheduling info is missing
In the case where multiple instances of the reaper notification
service exist, the scheduling information and the state update
notification might be picked up by different listeners. To make sure
that the state update notification ends up to the correct listener
we requeue the notification when the scheduling notification is
missing.