Persist notification events
With this change, the notifications received from nova are being
persisted in the db. In the scenario where we have multiple reaper
workers, storing in memory is not enough. Up till now, each worker,
upon receiving the state update event, it checked its in-memory
keystore to retrieve the scheduling info. This means only the worker
that initially received the scheduling info could handle the state
update for an instance. Now all workers will be able to trigger the
reaper and handle every state update event received.