Fix resource reservation
Trying to keep the same codepath for both watermark and standard
modes is proving to be a bit more difficult than expected. This
change addresses an issue with the resource reservations and the
logic we iterate over the hosts. The watermark mode is best effort
and we need to gather the requested resources even if we cannot find
all of them. On the other hand, in the standard mode we only preempt
servers if all the requested resources are found. Reservations should
be considered only for the standard mode, because in the watermark
mode we loop over the same once at most.