Fix aggregates check
Requested aggregates should not be a subset of watched aggregates.
The resulting aggregates set should be the intersection of the two
sets. If the watched and requested aggregates have no common elements
then aardvark resets the instance to error.

The check was in the job_handler but in order to include such issues
in the action's fault reason, we are moving the check to the reaper.